GatewayServis/requirements.txt
wheel
nats-python
asyncio-nats-client
opencv-contrib-python
cherrypy
requests
imutils
ws4py